Remade Save/Load System
Saves and loads are now done with a binary formatter, which saves 78% more storage space, and reduces loading time by 99.8%. To make things simple, I renamed the new weight files from weightpersistence.dat to WeightSave.dat. The new format will be used from now on for all of my networks, but the PricePredict one will continue to use the old format due to the low complexity. There are two downsides to the upgrade though. The first is that since the weights are binary, you can't easily read or edit it. And the second is that the entire array is saved, but no more. That means a second file, called WeightSaveMeta.meta, now carries all of the additional information.